Updated name translation and name similarity functions and modified the http protocol selection mechanism.
