+++ examples/entities_example.js
@@ -11,7 +11,7 @@ parser.addArgument(["--key"], {help: "Rosette API key", required: true});
 parser.addArgument(["--url"], {help: "Rosette API alt-url", required: false});
 var args = parser.parseArgs();
 
-var api = new Api(args.key);
+var api = new Api(args.key, args.url);
 var endpoint = "entities";
 
 api.parameters.content = "Bill Murray will appear in new Ghostbusters film: Dr. Peter Venkman was spotted filming a cameo in Boston this… http://dlvr.it/BnsFfS";
