--- examples/info_examples.js
@@ -1,22 +1,0 @@
-"use strict";
-
-var Api = require("../lib/Api");
-var ArgumentParser = require("argparse").ArgumentParser;
-
-var parser = new ArgumentParser({
-  addHelp: true,
-  description: "Get information about Rosette API"
-});
-parser.addArgument(["--key"], {help: "Rosette API key", required: true});
-parser.addArgument(["--url"], {help: "Rosette API alt-url", required: false});
-var args = parser.parseArgs();
-var api = new Api(args.key);
-var endpoint = "info";
-
-api.rosette(endpoint, function(err, res){
-	if(err){
-		console.log(err);
-	} else {
-		console.log(JSON.stringify(res, null, 2));
-	}
-});
