+++ examples/relationships_example.js
@@ -11,7 +11,7 @@ parser.addArgument(["--key"], {help: "Rosette API key", required: true});
 parser.addArgument(["--url"], {help: "Rosette API alt-url", required: false});
 var args = parser.parseArgs();
 
-var api = new Api(args.key);
+var api = new Api(args.key, args.url);
 var endpoint = "relationships";
 
 var relationships_text_data = "Bill Murray is in the new Ghostbusters film!";
