+++ lib/Api.js
@@ -18,8 +18,8 @@ var RosetteException = require("./rosetteExceptions");
 var paramObj = require("./parameters");
 
 var relationships = require("./relationships");
-var matchedName = require("./matchedName");
-var translatedName = require("./translatedName");
+var matchedName = require("./nameSimilarity");
+var translatedName = require("./nameTranslation");
 var sentiment = require("./sentiment");
 var categories = require("./categories");
 var entities = require("./entities");
@@ -67,9 +67,10 @@ function Api(userKey, serviceURL) {
      * @desc URL of the API
      * @default
      */
-    this.serviceURL = "https://api.rosette.com/rest/v1/";
     if (serviceURL) {
-        this.serviceURL = serviceURL;
+       this.serviceURL = serviceURL;
+    } else {
+        this.serviceURL = "http://cnode04:9119/rest/v1/";
     }
 
     this.versionChecked = false;
