+++ lib/categories.js
@@ -60,7 +60,7 @@ categories.prototype.getResults = function(parameters, userKey, serviceURL, call
         // configure URL
         var urlParts = URL.parse(serviceURL + "categories");
         var protocol = https;
-        if (urlParts.protocol === "http") {
+        if (urlParts.protocol === "http:") {
             protocol = http;
         }
