+++ lib/checkVersion.js
@@ -45,12 +45,11 @@ function checkVersion() {
  * @param {function} callback - Callback function to be exectuted after the function to which it is passed is complete
  */
 checkVersion.prototype.check = function(userKey, serviceURL, callback) {
-
     // configure URL
     var urlParts = URL.parse(serviceURL + "info?clientVersion=" + BINDING_VERSION);
     var protocol = https;
-    if (urlParts.protocol === "http") {
-        protocol = http;
+    if (urlParts.protocol === "http:") {
+       var protocol = http;
     }
 
     var headers = {
@@ -68,7 +67,7 @@ checkVersion.prototype.check = function(userKey, serviceURL, callback) {
         path: urlParts.path,
         method: 'POST',
         headers: headers,
-        agent: false
+        agent: false, 
     };
 
     if (urlParts.port) {
