+++ lib/entities.js
@@ -65,7 +65,7 @@ entities.prototype.getResults = function(parameters, userKey, serviceURL, callba
         }
 
         var protocol = https;
-        if (urlParts.protocol === "http") {
+        if (urlParts.protocol === "http:") {
             protocol = http;
         }
