+++ lib/info.js
@@ -50,8 +50,8 @@ info.prototype.getResults = function(parameters, userKey, serviceURL, callback) 
     // configure URL
     var urlParts = URL.parse(serviceURL + "info");
     var protocol = https;
-    if (urlParts.protocol === "http") {
-        protocol = http;
+    if (urlParts.protocol === "http:") {
+       var protocol = http;
     }
 
     var headers = {
@@ -65,11 +65,13 @@ info.prototype.getResults = function(parameters, userKey, serviceURL, callback) 
     var result = new Buffer("");
 
     var options = {
-        hostname: urlParts.hostname,
+        hostname: 'jugmaster',
         path: urlParts.path,
         method: 'GET',
         headers: headers,
-        agent: false
+        agent: false,
+        port: 8181,
+
     };
 
     if (urlParts.port) {
