+++ lib/language.js
@@ -60,7 +60,7 @@ language.prototype.getResults = function(parameters, userKey, serviceURL, callba
         // configure URL
         var urlParts = URL.parse(serviceURL + "language");
         var protocol = https;
-        if (urlParts.protocol === "http") {
+        if (urlParts.protocol === "http:") {
             protocol = http;
         }
