+++ lib/morphology.js
@@ -62,7 +62,7 @@ morphology.prototype.getResults = function(parameters, userKey, serviceURL, call
         // configure URL
         var urlParts = URL.parse(serviceURL + "morphology" + "/" + parameters.loadParams().morphology);
         var protocol = https;
-        if (urlParts.protocol === "http") {
+        if (urlParts.protocol === "http:") {
             protocol = http;
         }
