+++ lib/parameters.js
@@ -29,7 +29,6 @@ function parameters() {
     this.content = null;
     this.contentUri = null;
     this.contentType = null;
-    this.unit = null;
     this.language = null;
     this.documentFile = null;
 
@@ -72,7 +71,6 @@ parameters.prototype.loadParams = function() {
         "content": this.content,
         "contentUri": this.contentUri,
         "contentType": this.contentType,
-        "unit": this.unit,
         "language": this.language,
         "documentFile": this.documentFile,
         "name1": this.name1,
@@ -111,7 +109,6 @@ parameters.prototype.loadFile = function(filePath) {
 
     this.content = str;
     this.contentType = RosetteConstants.dataFormat.UNSPECIFIED;
-    this.unit = RosetteConstants.inputUnit.DOC;
 }
 
 module.exports = parameters;
