+++ lib/ping.js
@@ -50,7 +50,7 @@ ping.prototype.getResults = function(parameters, userKey, serviceURL, callback) 
     // configure URL
     var urlParts = URL.parse(serviceURL + "ping");
     var protocol = https;
-    if (urlParts.protocol === "http") {
+    if (urlParts.protocol === "http:") {
         protocol = http;
     }
