+++ lib/relationships.js
@@ -61,7 +61,7 @@ relationships.prototype.getResults = function(parameters, userKey, serviceURL, c
     // configure URL
     var urlParts = URL.parse(serviceURL + "relationships");
     var protocol = https;
-    if (urlParts.protocol === "http") {
+    if (urlParts.protocol === "http:") {
         protocol = http;
     }
