+++ lib/sentences.js
@@ -60,7 +60,7 @@ sentences.prototype.getResults = function(parameters, userKey, serviceURL, callb
         // configure URL
         var urlParts = URL.parse(serviceURL + "sentences");
         var protocol = https;
-        if (urlParts.protocol === "http") {
+        if (urlParts.protocol === "http:") {
             protocol = http;
         }
