+++ lib/sentiment.js
@@ -60,7 +60,7 @@ sentiment.prototype.getResults = function(parameters, userKey, serviceURL, callb
         // configure URL
         var urlParts = URL.parse(serviceURL + "sentiment");
         var protocol = https;
-        if (urlParts.protocol === "http") {
+        if (urlParts.protocol === "http:") {
             protocol = http;
         }
