+++ lib/tokens.js
@@ -60,7 +60,7 @@ tokens.prototype.getResults = function(parameters, userKey, serviceURL, callback
         // configure URL
         var urlParts = URL.parse(serviceURL + "tokens");
         var protocol = https;
-        if (urlParts.protocol === "http") {
+        if (urlParts.protocol === "http:") {
             protocol = http;
         }
